refactor: extract runtime command registry
