+++ packages/runtime-core/src/index.ts
@@ -4,6 +4,7 @@ import { dirname, isAbsolute, join, normalize, relative, resolve, sep } from "no
 
 export * from "./workspace-policy.js"
 export * from "./sandbox-datamachine-tool-policy.js"
+export * from "./command-registry.js"
 
 export type RuntimeBackendKind = "wordpress-playground" | (string & {})
 
@@ -15,244 +16,6 @@ export const RUNTIME_REFERENCE_MANIFEST_SCHEMA = "wp-codebox/runtime-reference-m
 export const RUNTIME_REPLAY_REFERENCE_INDEX_SCHEMA = "wp-codebox/runtime-replay-reference-index/v1" as const
 export const RUNTIME_ACTION_OBSERVATION_SCHEMA = "wp-codebox/runtime-action-observation/v1" as const
 export const BROWSER_INTERACTION_SCRIPT_SCHEMA = "wp-codebox/browser-interaction-script/v1" as const
-
-export type CommandHandlerBinding =
-  | { kind: "playground"; method: string }
-  | { kind: "recipe-alias"; command: string }
-
-export interface CommandDefinition {
-  id: string
-  description: string
-  acceptedArgs: Array<{
-    name: string
-    description: string
-    required?: boolean
-    repeatable?: boolean
-    format?: string
-  }>
-  outputShape: string
-  policyRequirement: string
-  recipe: boolean
-  handler: CommandHandlerBinding
-}
-
-export const commandRegistry = [
-  {
-    id: "inspect-mounted-inputs",
-    description: "List mounted input entries visible inside the Playground runtime.",
-    acceptedArgs: [],
-    outputShape: "JSON array of mounted input descriptors.",
-    policyRequirement: "Runtime policy commands must include inspect-mounted-inputs.",
-    recipe: true,
-    handler: { kind: "playground", method: "inspectMountedInputs" },
-  },
-  {
-    id: "wordpress.run-php",
-    description: "Run PHP against WordPress, bootstrapping wp-load.php unless bootstrap=none is supplied.",
-    acceptedArgs: [
-      { name: "code", description: "Inline PHP code to run.", format: "PHP string" },
-      { name: "code-file", description: "Path to a PHP file whose contents should run.", format: "path" },
-      { name: "bootstrap", description: "Use bootstrap=none to skip wp-load.php.", format: "wordpress|none" },
-    ],
-    outputShape: "Raw command stdout from the PHP snippet.",
-    policyRequirement: "Runtime policy commands must include wordpress.run-php.",
-    recipe: true,
-    handler: { kind: "playground", method: "runPhp" },
-  },
-  {
-    id: "wordpress.wp-cli",
-    description: "Run a WP-CLI command inside the same disposable WordPress runtime.",
-    acceptedArgs: [
-      { name: "command", description: "WP-CLI command line, with or without the leading wp token.", required: true, format: "string" },
-    ],
-    outputShape: "Raw WP-CLI stdout.",
-    policyRequirement: "Runtime policy commands must include wordpress.wp-cli.",
-    recipe: true,
-    handler: { kind: "playground", method: "runWpCli" },
-  },
-  {
-    id: "wordpress.ability",
-    description: "Execute a registered WordPress Ability in the sandbox.",
-    acceptedArgs: [
-      { name: "name", description: "Ability name to execute.", required: true, format: "string" },
-      { name: "input", description: "Ability input payload.", format: "JSON object" },
-    ],
-    outputShape: "JSON object with command, name, input, and result fields.",
-    policyRequirement: "Runtime policy commands must include wordpress.ability.",
-    recipe: true,
-    handler: { kind: "playground", method: "runAbility" },
-  },
-  {
-    id: "wordpress.bench",
-    description: "Run plugin benchmark workloads and emit a normalized benchmark results envelope.",
-    acceptedArgs: [
-      { name: "component-id", description: "Component id for the benchmark results envelope.", format: "string" },
-      { name: "plugin-slug", description: "Plugin slug containing tests/bench workloads.", required: true, format: "slug" },
-      { name: "iterations", description: "Measured iterations per workload.", format: "positive integer" },
-      { name: "warmup", description: "Warmup iterations before measurement.", format: "non-negative integer" },
-      { name: "dependency-slugs", description: "Comma-separated plugin dependency slugs to load.", format: "comma-separated slugs" },
-      { name: "env-json", description: "Benchmark environment object.", format: "JSON object" },
-      { name: "workloads-json", description: "Explicit workload list.", format: "JSON array" },
-    ],
-    outputShape: "Benchmark results JSON envelope with component_id, iterations, and scenarios.",
-    policyRequirement: "Runtime policy commands must include wordpress.bench.",
-    recipe: true,
-    handler: { kind: "playground", method: "runBench" },
-  },
-  {
-    id: "wordpress.phpunit",
-    description: "Run plugin PHPUnit tests with normalized diagnostics and test-result artifact capture.",
-    acceptedArgs: [
-      { name: "plugin-slug", description: "Plugin slug under wp-content/plugins.", format: "slug" },
-      { name: "code", description: "Inline override PHP runner code.", format: "PHP string" },
-      { name: "code-file", description: "Path to override PHP runner code.", format: "path" },
-      { name: "autoload-file", description: "PHPUnit/vendor autoload path inside the sandbox.", format: "sandbox path" },
-      { name: "tests-dir", description: "WP PHPUnit tests directory inside the sandbox.", format: "sandbox path" },
-      { name: "phpunit-xml", description: "phpunit.xml path inside the plugin.", format: "path" },
-      { name: "test-file", description: "Single test file to run.", format: "path" },
-      { name: "changed-tests-json", description: "Changed test files for diagnostics.", format: "JSON array" },
-      { name: "env-json", description: "PHPUnit environment values.", format: "JSON object" },
-      { name: "wp-config-defines-json", description: "wp-config.php constants for the run.", format: "JSON object" },
-      { name: "dependency-mounts", description: "Comma-separated mounted dependency paths.", format: "comma-separated sandbox paths" },
-      { name: "multisite", description: "Run as multisite.", format: "boolean" },
-    ],
-    outputShape: "Raw PHPUnit runner JSON/log output plus normalized test-results artifact when artifacts are collected.",
-    policyRequirement: "Runtime policy commands must include wordpress.phpunit.",
-    recipe: true,
-    handler: { kind: "playground", method: "runPhpunit" },
-  },
-  {
-    id: "wordpress.plugin-check",
-    description: "Run the official WordPress Plugin Check plugin against a mounted plugin and emit normalized findings.",
-    acceptedArgs: [
-      { name: "plugin-slug", description: "Plugin slug under wp-content/plugins to validate.", required: true, format: "slug" },
-      { name: "checks", description: "Optional comma-separated official Plugin Check slugs to run; omit to run the default suite.", format: "comma-separated check slugs" },
-    ],
-    outputShape: "wp-codebox/plugin-check/v1 JSON with command, target plugin, exit code/status, summary counts, and findings; raw and normalized outputs are captured in artifacts.",
-    policyRequirement: "Runtime policy commands must include wordpress.plugin-check.",
-    recipe: true,
-    handler: { kind: "playground", method: "runPluginCheck" },
-  },
-  {
-    id: "wordpress.core-phpunit",
-    description: "Run WordPress core PHPUnit tests with normalized diagnostics and test-result artifact capture. PRECONDITION: the mounted wordpress-develop checkout MUST already have Composer dev dependencies installed (PHPUnit + yoast/phpunit-polyfills under vendor/) before mounting, because core's tests/phpunit/includes/bootstrap.php die()s without them. Run `composer install` (or `composer update -W`) in the checkout first, or mount one that already has vendor/. When the toolchain is missing the command now fails with a clear structured error instead of crashing silently.",
-    acceptedArgs: [
-      { name: "core-root", description: "WordPress develop checkout root inside the sandbox. Must contain vendor/ with Composer dev dependencies (PHPUnit + yoast/phpunit-polyfills) installed before mounting.", format: "sandbox path" },
-      { name: "tests-dir", description: "Core tests directory inside the sandbox (expects includes/bootstrap.php under it).", format: "sandbox path" },
-      { name: "phpunit-xml", description: "phpunit.xml path.", format: "path" },
-      { name: "test-file", description: "Single test file to run.", format: "path" },
-      { name: "changed-tests-json", description: "Changed test files for diagnostics.", format: "JSON array" },
-      { name: "autoload-file", description: "Autoload path inside the sandbox (typically <core-root>/vendor/autoload.php from a completed composer install).", format: "sandbox path" },
-      { name: "wp-config-defines-json", description: "wp-config.php constants for the run.", format: "JSON object" },
-      { name: "multisite", description: "Run as multisite.", format: "boolean" },
-    ],
-    outputShape: "Raw PHPUnit runner JSON/log output plus normalized test-results artifact when artifacts are collected.",
-    policyRequirement: "Runtime policy commands must include wordpress.core-phpunit.",
-    recipe: true,
-    handler: { kind: "playground", method: "runCorePhpunit" },
-  },
-  {
-    id: "wordpress.theme-check",
-    description: "Run Theme Check against a mounted WordPress theme inside the disposable Playground runtime.",
-    acceptedArgs: [
-      { name: "theme", description: "Theme slug under wp-content/themes.", required: true, format: "slug" },
-    ],
-    outputShape: "Normalized Theme Check JSON plus files/theme-check raw and normalized artifacts.",
-    policyRequirement: "Runtime policy commands must include wordpress.theme-check.",
-    recipe: true,
-    handler: { kind: "playground", method: "runThemeCheck" },
-  },
-  {
-    id: "wordpress.browser-probe",
-    description: "Open the live Playground preview in Playwright and capture generic browser replay/audit evidence artifacts.",
-    acceptedArgs: [
-      { name: "url", description: "Preview path or absolute URL to visit.", required: true, format: "path or URL" },
-      { name: "wait-for", description: "Navigation wait condition.", format: "domcontentloaded|load|networkidle|selector:<selector>|duration" },
-      { name: "duration", description: "Extra capture duration, or wait time when wait-for=duration.", format: "duration, e.g. 2s or 500ms" },
-      { name: "script", description: "Optional page-side JavaScript to evaluate after navigation and before final capture.", format: "JavaScript function body" },
-      { name: "capture", description: "Comma-separated artifacts to capture.", format: "console,errors,html,network,performance,memory,screenshot" },
-      { name: "repeat", description: "Optional repeated probe iterations for leak-oriented recipes.", format: "positive integer" },
-      { name: "reset-between", description: "Requested reset mode between repeated probe iterations.", format: "none|reload|new-page" },
-    ],
-    outputShape: "JSON summary plus files/browser/console.jsonl, errors.jsonl, network.jsonl, performance.json, memory.json, checkpoints.jsonl, snapshot.html, summary.json, and screenshot.png when captured.",
-    policyRequirement: "Runtime policy commands must include wordpress.browser-probe.",
-    recipe: true,
-    handler: { kind: "playground", method: "runBrowserProbe" },
-  },
-  {
-    id: "wordpress.browser-actions",
-    description: "Drive the live Playground preview with an ordered interaction script and capture replay/audit evidence artifacts, including per-step results and machine-readable assertions.",
-    acceptedArgs: [
-      { name: "url", description: "Initial preview path or absolute URL to visit when the script omits an initial navigate step.", format: "path or URL" },
-      { name: "steps-json", description: "Ordered interaction script: navigate, click, fill, type, press, drag, hover, select, waitFor, evaluate, expect, screenshot, and capture steps.", format: "JSON array (inline or @<path>)" },
-      { name: "actions-json", description: "Back-compat alias for steps-json accepting the legacy navigate/click/fill/press/wait/capture action shape.", format: "JSON array" },
-      { name: "step-timeout", description: "Per-step timeout applied to each interaction step.", format: "duration, e.g. 5s or 500ms" },
-      { name: "timeout", description: "Total-script timeout bounding the whole interaction run.", format: "duration, e.g. 30s or 1500ms" },
-      { name: "capture", description: "Comma-separated artifacts to capture after interactions.", format: "steps,console,errors,html,network,screenshot" },
-    ],
-    outputShape: "JSON summary plus files/browser/steps.jsonl, action-summary.json (with assertions pass/fail), named screenshots, and optional console/errors/network/html/screenshot artifacts.",
-    policyRequirement: "Runtime policy commands must include wordpress.browser-actions. The evaluate step additionally requires wordpress.browser-actions.evaluate.",
-    recipe: true,
-    handler: { kind: "playground", method: "runBrowserActions" },
-  },
-  {
-    id: "wordpress.browser-actions.evaluate",
-    description: "Policy capability gating arbitrary page-side JavaScript (the evaluate step) inside wordpress.browser-actions. Non-JS interaction steps do not require this capability.",
-    acceptedArgs: [],
-    outputShape: "Policy-only capability; not directly executable. Grant alongside wordpress.browser-actions to permit evaluate steps.",
-    policyRequirement: "Runtime policy commands must include wordpress.browser-actions.evaluate to run evaluate steps.",
-    recipe: false,
-    handler: { kind: "recipe-alias", command: "wordpress.browser-actions" },
-  },
-  {
-    id: "wp-codebox.agent-runtime-probe",
-    description: "Recipe-only probe that boots Agents API, Data Machine, and Data Machine Code and verifies the stack loads.",
-    acceptedArgs: [
-      { name: "provider-plugin-slugs", description: "Comma-separated provider plugin slugs already mounted by recipe inputs.", format: "comma-separated slugs" },
-    ],
-    outputShape: "JSON probe result emitted by the sandbox PHP runner.",
-    policyRequirement: "Recipe policy maps this helper to wordpress.run-php.",
-    recipe: true,
-    handler: { kind: "recipe-alias", command: "wordpress.run-php" },
-  },
-  {
-    id: "wp-codebox.agent-sandbox-run",
-    description: "Recipe-only helper that runs one natural-language task through the sandboxed agent stack.",
-    acceptedArgs: [
-      { name: "task", description: "Task prompt for the sandbox agent.", required: true, format: "string" },
-      { name: "agent", description: "Agent slug.", format: "string" },
-      { name: "mode", description: "Agent mode.", format: "string" },
-      { name: "provider", description: "AI provider id.", format: "string" },
-      { name: "model", description: "Model id.", format: "string" },
-      { name: "session-id", description: "Conversation session id.", format: "string" },
-      { name: "max-turns", description: "Maximum agent loop turns.", format: "positive integer" },
-      { name: "timeout-seconds", description: "Maximum wall-clock seconds for the sandbox agent PHP task.", format: "positive integer" },
-      { name: "provider-plugin-slugs", description: "Comma-separated provider plugin slugs already mounted by recipe inputs.", format: "comma-separated slugs" },
-      { name: "code", description: "Inline PHP runner override for operator/debug use.", format: "PHP string" },
-      { name: "code-file", description: "Path to PHP runner override for operator/debug use.", format: "path" },
-    ],
-    outputShape: "JSON agent run result emitted by the sandbox PHP runner.",
-    policyRequirement: "Recipe policy maps this helper to wordpress.run-php.",
-    recipe: true,
-    handler: { kind: "recipe-alias", command: "wordpress.run-php" },
-  },
-] as const satisfies readonly CommandDefinition[]
-
-export type CommandId = typeof commandRegistry[number]["id"]
-export type PlaygroundRuntimeCommandDefinition = Extract<typeof commandRegistry[number], { handler: { kind: "playground" } }>
-export type PlaygroundRuntimeCommandId = PlaygroundRuntimeCommandDefinition["id"]
-
-export function getCommandDefinition(command: string): CommandDefinition | undefined {
-  return commandRegistry.find((definition) => definition.id === command)
-}
-
-export function runtimeCommandDefinitions(): CommandDefinition[] {
-  return commandRegistry.filter((definition) => definition.handler.kind === "playground")
-}
-
-export function recipeCommandDefinitions(): CommandDefinition[] {
-  return commandRegistry.filter((definition) => definition.recipe)
-}
 
 export const RUNTIME_EPISODE_TRACE_JSON_SCHEMA = {
   $id: RUNTIME_EPISODE_TRACE_SCHEMA,
